+++ R/modelStudio.R
@@ -176,14 +176,14 @@ modelStudio.explainer <- function(explainer,
 
   } else if (is.null(dim(new_observation))) {
     warning("`new_observation` argument is not a data.frame nor a matrix, coerced to data.frame\n")
-    new_observation <- as.data.frame(new_observation)
+    new_observation <- as.data.frame(new_observation, stringsAsFactors=TRUE)
 
   } else if (is.null(rownames(new_observation))) {
     rownames(new_observation) <- 1:nrow(new_observation)
   }
 
   check_single_prediction <- try(predict_function(model, new_observation[1,, drop = FALSE]), silent = TRUE)
-  if (class(check_single_prediction)[1] == "try-error") {
+  if ("try-error" %in% class(check_single_prediction)) {
     stop("`explainer$predict_function` returns an error when executed on `new_observation[1,, drop = FALSE]` \n")
   }
   #:#
@@ -345,7 +345,8 @@ modelStudio.explainer <- function(explainer,
   between <- " - "
   if (is.null(new_observation_y)) new_observation_y <- between <- ""
   drop_down_data <- as.data.frame(cbind(rownames(obs_data),
-                                        paste0(rownames(obs_data), between, new_observation_y)))
+                                        paste0(rownames(obs_data), between, new_observation_y)),
+                                  stringsAsFactors=TRUE)
   colnames(drop_down_data) <- c("id", "text")
 
   # prepare footer text and ms title