+++ R/ms_update_observations.R
@@ -114,7 +114,7 @@ ms_update_observations <- function(object,
   }
 
   check_single_prediction <- try(predict_function(model, new_observation[1,, drop = FALSE]), silent = TRUE)
-  if (class(check_single_prediction)[1] == "try-error") {
+  if ("try-error" %in% class(check_single_prediction)) {
     stop("`predict_function` returns an error when executed on `new_observation[1,, drop = FALSE]` \n")
   }